fix(parser): Improve types for parser.py (#419)
When attempting to use the parse function we noticed we were getting
error messages like:

    Value of type variable "Model" of "parse" cannot be "Type[UserModel]"

Further investigation (using the example shown at
https://awslabs.github.io/aws-lambda-powertools-python/latest/utilities/parser/#envelopes)
showed that the types were incorrect for the example. The function
signatures were expecting an instance of `Model`, however the example
(and code) seemeed to expect a `Type[Model]`.

Similarly the signatures were expecting an optional instance of an
Envelope, but the code allows for either an instance or a Type.

Also, added a `TypeVar` for the decorator, specifying that the
event_parser` decorator returns the same type as the handler used.

Finally updated documentation on use of parser.